Rename RAT category to RMM
The acronym for Remote Administration Tool (RAT) is often mixed up with the one for Remote Access Trojan (RAT). The first one is a legitimate software to manage remote systems, whereas the second type is clearly malicious and would be part of the `hacking` category.
To avoid possible confusion, rename RAT to RMM (Remote Monitoring & Management).